feat: general quick rewrite
* FileGenerator has been renamed to File and moved to base folder
* Added mandatory PIID (Plugin Identifier & ID) to plugins
