+++ .vscode/tasks.json
@@ -12,6 +12,18 @@
       "options": {
         "cwd": "${workspaceFolder}/docs"
       }
+    },
+    {
+      "label": "Autobuild API docs",
+      "type": "shell",
+      "command": "poetry",
+      "args": ["run", "sphinx-apidoc", "${cwd}/fabricius", "-o", "${cwd}/docs/source/api", "--force"],
+      "group": {
+        "kind": "build"
+      },
+      "options": {
+        "cwd": "${workspaceFolder}"
+      }
     }
   ]
 }
