+++ docs/source/api/fabricius.generator.rst
@@ -4,34 +4,10 @@ fabricius.generator package
 Submodules
 ----------
 
-fabricius.generator.errors module
----------------------------------
-
-.. automodule:: fabricius.generator.errors
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fabricius.generator.file module
--------------------------------
-
-.. automodule:: fabricius.generator.file
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 fabricius.generator.generator module
 ------------------------------------
 
 .. automodule:: fabricius.generator.generator
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fabricius.generator.renderer module
------------------------------------
-
-.. automodule:: fabricius.generator.renderer
    :members:
    :undoc-members:
    :show-inheritance:
