+++ docs/source/api/fabricius.rst
@@ -29,10 +29,26 @@ fabricius.errors module
    :undoc-members:
    :show-inheritance:
 
+fabricius.file module
+---------------------
+
+.. automodule:: fabricius.file
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 fabricius.interfaces module
 ---------------------------
 
 .. automodule:: fabricius.interfaces
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fabricius.renderer module
+-------------------------
+
+.. automodule:: fabricius.renderer
    :members:
    :undoc-members:
    :show-inheritance:
