+++ docs/source/conf.py
@@ -74,5 +74,5 @@ coverage_show_missing_items = True
 
 # -- Options for intersphinx extension ---------------------------------------
 intersphinx_mapping = {
-    "python": ("https://docs.python.org/3", None),
+    "py": ("https://docs.python.org/3", None),
 }
