+++ docs/source/guides/guide_project_rendering.rst
@@ -1,0 +1,6 @@
+Guide : Rendering a template project
+====================================
+
+Hi! Looks like you found a work in progress, how about you check this out later? :)
+
+No deadlines yet...
